backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic[email]>=2.10.0
langchain
boto3
python-dotenv
supabase
openai
pytest>=8.0.0
httpx>=0.27.0
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.10
PyMuPDF
python-pptx
python-multipart